Implement add docs
