+++ src/marqo/tensor_search/tensor_search.py
@@ -31,54 +31,58 @@ Notes on search behaviour with caching and searchable attributes:
 
 """
 import copy
+import functools
 import json
+import pprint
+import threading
+import typing
+import uuid
 from collections import defaultdict
 from contextlib import ExitStack
 from timeit import default_timer as timer
-import functools
-import pprint
-import typing
-from marqo.tensor_search.models.private_models import ModelAuth
-import uuid
 from typing import List, Optional, Union, Iterable, Sequence, Dict, Any, Tuple
-from marqo.tensor_search.models.add_docs_objects import AddDocsParams
+
 import numpy as np
+import psutil
+import torch.cuda
 from PIL import Image
+
 import marqo.config as config
-from marqo.tensor_search.models.delete_docs_objects import MqDeleteDocsRequest
+from marqo import errors
+# We depend on _httprequests.py for now, but this may be replaced in the future, as
+# _httprequests.py is designed for the client
+from marqo._httprequests import HttpRequests
+from marqo.config import Config
+from marqo.core.models.marqo_index import IndexType, MarqoIndex, FieldType
+from marqo.core.typed_vespa_index import TypedVespaIndex
+from marqo.s2_inference import errors as s2_inference_errors
+from marqo.s2_inference import s2_inference
+from marqo.s2_inference.clip_utils import _is_image
+from marqo.s2_inference.processing import image as image_processor
+from marqo.s2_inference.processing import text as text_processor
+from marqo.s2_inference.reranking import rerank
+from marqo.tensor_search import delete_docs
+from marqo.tensor_search import index_meta_cache
+from marqo.tensor_search import utils, backend, validation, configs, add_docs, filtering
 from marqo.tensor_search.enums import (
     Device, MediaType, MlModel, TensorField, SearchMethod, OpenSearchDataType,
     EnvVars
 )
 from marqo.tensor_search.enums import IndexSettingsField as NsField
-from marqo.tensor_search import utils, backend, validation, configs, add_docs, filtering
 from marqo.tensor_search.formatting import _clean_doc
-from marqo.tensor_search.index_meta_cache import get_cache, get_index_info
-from marqo.tensor_search import index_meta_cache
-from marqo.tensor_search.models.api_models import BulkSearchQuery, BulkSearchQueryEntity, ScoreModifier
-from marqo.tensor_search.models.search import Qidx, JHash, SearchContext, VectorisedJobs, VectorisedJobPointer
-from marqo.tensor_search.models.index_info import IndexInfo, get_model_properties_from_index_defaults
-from marqo.tensor_search.models.external_apis.abstract_classes import ExternalAuth
-from marqo.tensor_search.telemetry import RequestMetricsStore
 from marqo.tensor_search.health import generate_heath_check_response
-from marqo.tensor_search.utils import add_timing
-from marqo.tensor_search import delete_docs
-from marqo.s2_inference.processing import text as text_processor
-from marqo.s2_inference.processing import image as image_processor
-from marqo.s2_inference.clip_utils import _is_image
-from marqo.s2_inference.reranking import rerank
-from marqo.s2_inference import s2_inference
-import torch.cuda
-import psutil
-# We depend on _httprequests.py for now, but this may be replaced in the future, as
-# _httprequests.py is designed for the client
-from marqo._httprequests import HttpRequests
-from marqo.config import Config
-from marqo import errors
-from marqo.s2_inference import errors as s2_inference_errors
-import threading
-from dataclasses import replace
+from marqo.tensor_search.index_meta_cache import get_cache, get_index_info
+from marqo.tensor_search.models.add_docs_objects import AddDocsParams
+from marqo.tensor_search.models.api_models import BulkSearchQuery, BulkSearchQueryEntity, ScoreModifier
+from marqo.tensor_search.models.delete_docs_objects import MqDeleteDocsRequest
+from marqo.tensor_search.models.index_info import IndexInfo, get_model_properties_from_index_defaults
+from marqo.tensor_search.models.private_models import ModelAuth
+from marqo.tensor_search.models.search import Qidx, JHash, SearchContext, VectorisedJobs, VectorisedJobPointer
+from marqo.tensor_search.telemetry import RequestMetricsStore
 from marqo.tensor_search.tensor_search_logging import get_logger
+from marqo.tensor_search.utils import add_timing
+from marqo.vespa.models import VespaDocument
+from marqo.vespa.vespa_client import VespaClient
 
 logger = get_logger(__name__)
 
@@ -339,6 +343,24 @@ def _get_chunks_for_field(field_name: str, doc_id: str, doc):
 
 
 def add_documents(config: Config, add_docs_params: AddDocsParams):
+    try:
+        marqo_index = index_meta_cache.get_index(config=config, index_name=add_docs_params.index_name)
+    except errors.IndexNotFoundError:
+        raise errors.IndexNotFoundError(f"Cannot add documents to non-existent index {add_docs_params.index_name}")
+
+    if marqo_index.type == IndexType.Unstructured:
+        return _add_documents_unstructured(add_docs_params, marqo_index, config.vespa_client)
+    elif marqo_index.type == IndexType.Structured:
+        return _add_documents_structured(add_docs_params, marqo_index, config.vespa_client)
+    else:
+        raise errors.InternalError(f"Unknown index type {marqo_index.type}")
+
+
+def _add_documents_unstructured(add_docs_params: AddDocsParams, marqo_index: MarqoIndex, vespa_client: VespaClient):
+    pass
+
+
+def _add_documents_structured(add_docs_params: AddDocsParams, marqo_index: MarqoIndex, vespa_client: VespaClient):
     """
     Args:
         config: Config object
@@ -355,25 +377,11 @@ def add_documents(config: Config, add_docs_params: AddDocsParams):
     t0 = timer()
     bulk_parent_dicts = []
 
-    try:
-        index_info = backend.get_index_info(config=config, index_name=add_docs_params.index_name)
-    except errors.IndexNotFoundError:
-        raise errors.IndexNotFoundError(f"Cannot add documents to non-existent index {add_docs_params.index_name}")
-
     if len(add_docs_params.docs) == 0:
         raise errors.BadRequestError(message="Received empty add documents request")
 
     if add_docs_params.mappings is not None:
         validate_mappings = validation.validate_mappings(add_docs_params.mappings)
-
-    existing_fields = set(index_info.properties.keys())
-    new_fields = set()
-
-    # A dict to store the multimodal_fields and their (child_fields, opensearch_type)
-    # dict = {parent_field_1 : set((child_field_1, type), ),
-    #      =  parent_field_2 : set((child_field_1, type), )}
-    # Check backend to see the differences between multimodal_fields and new_fields
-    new_obj_fields = dict()
 
     unsuccessful_docs = []
     total_vectorise_time = 0
@@ -381,43 +389,46 @@ def add_documents(config: Config, add_docs_params: AddDocsParams):
     image_repo = {}
 
     with ExitStack() as exit_stack:
-        if index_info.index_settings[NsField.index_defaults][NsField.treat_urls_and_pointers_as_images]:
+        tensor_fields = marqo_index.tensor_field_map.keys()
+        image_fields = marqo_index.image_pointer_field_map.keys()
+
+        if image_fields:
             with RequestMetricsStore.for_request().time(
-                "image_download.full_time",
-                lambda t: logger.debug(
-                    f"add_documents image download: took {t:.3f}ms to concurrently download "
-                    f"images for {batch_size} docs using {add_docs_params.image_download_thread_count} threads"
-                )
+                    "image_download.full_time",
+                    lambda t: logger.debug(
+                        f"add_documents image download: took {t:.3f}ms to concurrently download "
+                        f"images for {batch_size} docs using {add_docs_params.image_download_thread_count} threads"
+                    )
             ):
-                if add_docs_params.tensor_fields and '_id' in add_docs_params.tensor_fields:
-                    raise errors.BadRequestError(message="`_id` field cannot be a tensor field.")
+
+                if '_id' in image_fields:
+                    raise errors.BadRequestError(message="`_id` field cannot be an image pointer field.")
 
                 image_repo = exit_stack.enter_context(
                     add_docs.download_images(docs=add_docs_params.docs, thread_count=20,
-                                             tensor_fields=add_docs_params.tensor_fields
-                                             if add_docs_params.tensor_fields is not None else None,
-                                             non_tensor_fields=add_docs_params.non_tensor_fields + ['_id']
-                                             if add_docs_params.non_tensor_fields is not None else None,
+                                             tensor_fields=image_fields,
+                                             non_tensor_fields=None,
                                              image_download_headers=add_docs_params.image_download_headers)
                 )
 
         if add_docs_params.use_existing_tensors:
-            doc_ids = []
+            raise errors.MarqoWebError('use_existing_tensors is not currently supported with Vespa')
 
-            # Iterate through the list in reverse, only latest doc with dupe id gets added.
-            for i in range(len(add_docs_params.docs)-1, -1, -1):
-                if ("_id" in add_docs_params.docs[i]) and (add_docs_params.docs[i]["_id"] not in doc_ids):
-                    doc_ids.append(add_docs_params.docs[i]["_id"])
-            existing_docs = _get_documents_for_upsert(
-                config=config, index_name=add_docs_params.index_name, document_ids=doc_ids)
+            # TODO - implement this for Vespa after get doc is added
+            # doc_ids = []
+            #
+            # # Iterate through the list in reverse, only latest doc with dupe id gets added.
+            # for i in range(len(add_docs_params.docs) - 1, -1, -1):
+            #     if ("_id" in add_docs_params.docs[i]) and (add_docs_params.docs[i]["_id"] not in doc_ids):
+            #         doc_ids.append(add_docs_params.docs[i]["_id"])
+            # existing_docs = _get_documents_for_upsert(
+            #     config=config, index_name=add_docs_params.index_name, document_ids=doc_ids)
 
         for i, doc in enumerate(add_docs_params.docs):
 
-            indexing_instructions = {'index': {"_index": add_docs_params.index_name}}
             copied = copy.deepcopy(doc)
 
             document_is_valid = True
-            new_fields_from_doc = set()
 
             doc_id = None
             try:
@@ -437,38 +448,46 @@ def add_documents(config: Config, add_docs_params: AddDocsParams):
                 )
                 continue
 
-            indexing_instructions["index"]["_id"] = doc_id
-            if add_docs_params.use_existing_tensors:
-                matching_doc = [doc for doc in existing_docs["docs"] if doc["_id"] == doc_id]
-                # Should only have 1 result, as only 1 id matches
-                if len(matching_doc) == 1:
-                    existing_doc = matching_doc[0]
-                # When a request isn't sent to get matching docs, because the added docs don't
-                # have IDs:
-                elif len(matching_doc) == 0:
-                    existing_doc = {"found": False}
-                else:
-                    raise errors.InternalError(message= f"Upsert: found {len(matching_doc)} matching docs for {doc_id} when only 1 or 0 should have been found.")
+            # indexing_instructions["index"]["_id"] = doc_id
+            # if add_docs_params.use_existing_tensors:
+            #     matching_doc = [doc for doc in existing_docs["docs"] if doc["_id"] == doc_id]
+            #     # Should only have 1 result, as only 1 id matches
+            #     if len(matching_doc) == 1:
+            #         existing_doc = matching_doc[0]
+            #     # When a request isn't sent to get matching docs, because the added docs don't
+            #     # have IDs:
+            #     elif len(matching_doc) == 0:
+            #         existing_doc = {"found": False}
+            #     else:
+            #         raise errors.InternalError(
+            #             message=f"Upsert: found {len(matching_doc)} matching docs for {doc_id} when only 1 or 0 should have been found.")
+            #
+            # # Metadata can be calculated here at the doc level.
+            # # Only add chunk values which are string, boolean, numeric or dictionary.
+            # # Dictionary keys will be store in a list.
+            # chunk_values_for_filtering = {}
+            # for key, value in copied.items():
+            #     if not (isinstance(value, str) or isinstance(value, float)
+            #             or isinstance(value, bool) or isinstance(value, int)
+            #             or isinstance(value, list) or isinstance(value, dict)):
+            #         continue
+            #     chunk_values_for_filtering[key] = value
 
-            # Metadata can be calculated here at the doc level.
-            # Only add chunk values which are string, boolean, numeric or dictionary.
-            # Dictionary keys will be store in a list.
-            chunk_values_for_filtering = {}
-            for key, value in copied.items():
-                if not (isinstance(value, str) or isinstance(value, float)
-                        or isinstance(value, bool) or isinstance(value, int)
-                        or isinstance(value, list) or isinstance(value, dict)):
-                    continue
-                chunk_values_for_filtering[key] = value
-
-            chunks = []
-
+            fields_to_add = {}
+            fields_to_remove = []
             for field in copied:
+                marqo_field = marqo_index.field_map.get(field)
+                if not marqo_field:
+                    raise errors.InvalidArgError(
+                        f"Field {field} is not a valid field for structured index {add_docs_params.index_name}. "
+                        f"Valid fields are: {', '.join(marqo_index.field_map.keys())}"
+                    )
 
                 try:
                     field_content = validation.validate_field_content(
                         field_content=copied[field],
-                        is_non_tensor_field=not utils.is_tensor_field(field, add_docs_params.tensor_fields, add_docs_params.non_tensor_fields)
+                        is_non_tensor_field=not utils.is_tensor_field(field, add_docs_params.tensor_fields,
+                                                                      add_docs_params.non_tensor_fields)
                     )
                     if isinstance(field_content, dict):
                         field_content = validation.validate_dict(
@@ -484,23 +503,23 @@ def add_documents(config: Config, add_docs_params: AddDocsParams):
                     )
                     break
 
-                if (field not in existing_fields) and (not isinstance(field_content, dict)):
-                    new_fields_from_doc.add((field, _infer_opensearch_data_type(copied[field])))
-
-                # Don't process text/image fields when explicitly told not to.
-                if not utils.is_tensor_field(field, add_docs_params.tensor_fields, add_docs_params.non_tensor_fields):
+                # Proceed from here only for tensor fields
+                tensor_field = marqo_index.tensor_field_map.get(field)
+                if not tensor_field:
                     continue
 
                 # chunks generated by processing this field for this doc:
-                chunks_to_append = []
+                chunks = []
+                embeddings = []
                 # Check if content of this field changed. If no, skip all chunking and vectorisation
-                if (add_docs_params.use_existing_tensors
-                        and existing_doc["found"]
-                        and (field in existing_doc["_source"]) and (existing_doc["_source"][field] == field_content)):
-                    chunks_to_append = _get_chunks_for_field(field_name=field, doc_id=doc_id, doc=existing_doc)
+                # if (add_docs_params.use_existing_tensors
+                #         and existing_doc["found"]
+                #         and (field in existing_doc["_source"]) and (existing_doc["_source"][field] == field_content)):
+                #     chunks_to_append = _get_chunks_for_field(field_name=field, doc_id=doc_id, doc=existing_doc)
 
                 # Chunk and vectorise, since content changed.
-                elif isinstance(field_content, (str, Image.Image)):
+                # elif isinstance(field_content, (str, Image.Image)):
+                if isinstance(field_content, (str, Image.Image)):
 
                     # TODO: better/consistent handling of a no-op for processing (but still vectorize)
 
@@ -513,25 +532,27 @@ def add_documents(config: Config, add_docs_params: AddDocsParams):
 
                     if isinstance(field_content, str) and not _is_image(field_content):
                         # text processing pipeline:
-                        split_by = index_info.index_settings[NsField.index_defaults][NsField.text_preprocessing][
-                            NsField.split_method]
-                        split_length = index_info.index_settings[NsField.index_defaults][NsField.text_preprocessing][
-                            NsField.split_length]
-                        split_overlap = index_info.index_settings[NsField.index_defaults][NsField.text_preprocessing][
-                            NsField.split_overlap]
+                        split_by = marqo_index.text_preprocessing.split_method.value
+                        split_length = marqo_index.text_preprocessing.split_length
+                        split_overlap = marqo_index.text_preprocessing.split_overlap
                         content_chunks = text_processor.split_text(field_content, split_by=split_by,
-                                                                   split_length=split_length, split_overlap=split_overlap)
+                                                                   split_length=split_length,
+                                                                   split_overlap=split_overlap)
                         text_chunks = content_chunks
                     else:
                         # TODO put the logic for getting field parameters into a function and add per field options
-                        image_method = index_info.index_settings[NsField.index_defaults][NsField.image_preprocessing][
-                            NsField.patch_method]
+                        image_method = marqo_index.image_preprocessing.patch_method
+
                         # the chunk_image contains the no-op logic as of now - method = None will be a no-op
                         try:
                             # in the future, if we have different chunking methods, make sure we catch possible
                             # errors of different types generated here, too.
-                            if isinstance(field_content, str) and index_info.index_settings[NsField.index_defaults][
-                                NsField.treat_urls_and_pointers_as_images]:
+                            if isinstance(field_content, str) and \
+                                    (marqo_index.type == IndexType.Unstructured and
+                                     marqo_index.treat_urls_and_pointers_as_images or
+                                     marqo_index.type == IndexType.Structured and
+                                     field in marqo_index.image_pointer_field_map):
+
                                 if not isinstance(image_repo[field_content], Exception):
                                     image_data = image_repo[field_content]
                                 else:
@@ -541,9 +562,9 @@ def add_documents(config: Config, add_docs_params: AddDocsParams):
                                     )
                             else:
                                 image_data = field_content
-                            if image_method not in [None, 'none', '', "None", ' ']:
+                            if image_method is not None:
                                 content_chunks, text_chunks = image_processor.chunk_image(
-                                    image_data, device=add_docs_params.device, method=image_method)
+                                    image_data, device=add_docs_params.device, method=image_method.value)
                             else:
                                 # if we are not chunking, then we set the chunks as 1-len lists
                                 # content_chunk is the PIL image
@@ -558,10 +579,7 @@ def add_documents(config: Config, add_docs_params: AddDocsParams):
                             )
                             break
 
-                    normalize_embeddings = index_info.index_settings[NsField.index_defaults][
-                        NsField.normalize_embeddings]
-                    infer_if_image = index_info.index_settings[NsField.index_defaults][
-                        NsField.treat_urls_and_pointers_as_images]
+                    normalize_embeddings = marqo_index.normalize_embeddings
 
                     try:
                         # in the future, if we have different underlying vectorising methods, make sure we catch possible
@@ -571,10 +589,10 @@ def add_documents(config: Config, add_docs_params: AddDocsParams):
                         start_time = timer()
                         with RequestMetricsStore.for_request().time(f"add_documents.create_vectors"):
                             vector_chunks = s2_inference.vectorise(
-                                model_name=index_info.model_name,
-                                model_properties=index_info.get_model_properties(), content=content_chunks,
+                                model_name=marqo_index.model.name,
+                                model_properties=marqo_index.model.properties, content=content_chunks,
                                 device=add_docs_params.device, normalize_embeddings=normalize_embeddings,
-                                infer=infer_if_image, model_auth=add_docs_params.model_auth
+                                infer=marqo_field.type == FieldType.ImagePointer, model_auth=add_docs_params.model_auth
                             )
 
                         end_time = timer()
@@ -603,79 +621,70 @@ def add_documents(config: Config, add_docs_params: AddDocsParams):
 
                     for text_chunk, vector_chunk in zip(text_chunks, vector_chunks):
                         # We do not put in metadata yet at this stage.
-                        chunks_to_append.append({
-                            TensorField.marqo_knn_field: vector_chunk,
-                            TensorField.field_content: text_chunk,
-                            TensorField.field_name: field
-                        })
+                        chunks.append(text_chunk)
+                        embeddings.append(vector_chunk)
 
                 elif isinstance(field_content, dict):
                     if add_docs_params.mappings[field]["type"] == "multimodal_combination":
                         (combo_chunk, combo_document_is_valid,
                          unsuccessful_doc_to_append, combo_vectorise_time_to_add,
                          new_fields_from_multimodal_combination) = vectorise_multimodal_combination_field(
-                                field, field_content, copied, i, doc_id, add_docs_params.device, index_info,
-                                image_repo, add_docs_params.mappings[field], model_auth=add_docs_params.model_auth)
+                            field, field_content, copied, i, doc_id, add_docs_params.device, marqo_index,
+                            image_repo, add_docs_params.mappings[field], model_auth=add_docs_params.model_auth)
                         total_vectorise_time = total_vectorise_time + combo_vectorise_time_to_add
                         if combo_document_is_valid is False:
                             unsuccessful_docs.append(unsuccessful_doc_to_append)
                             break
                         else:
-                            if field not in new_obj_fields:
-                                new_obj_fields[field] = set()
-                            new_obj_fields[field] = new_obj_fields[field].union(new_fields_from_multimodal_combination)
-                            # TODO: we may want to use chunks_to_append here to make it uniform with use_existing_tensors and normal vectorisation
-                            chunks.append({**combo_chunk, **chunk_values_for_filtering})
+                            # TODO - multimodal implementation
+                            # chunks.append({**combo_chunk, **chunk_values_for_filtering})
                             continue
 
                 # Add chunks_to_append along with doc metadata to total chunks
-                for chunk in chunks_to_append:
-                    chunks.append({**chunk, **chunk_values_for_filtering})
+                fields_to_add[tensor_field.chunk_field_name] = chunks
+                fields_to_add[tensor_field.embeddings_field_name] = embeddings
 
             if document_is_valid:
-                new_fields = new_fields.union(new_fields_from_doc)
-                copied[TensorField.chunks] = chunks
-                bulk_parent_dicts.append(indexing_instructions)
+                copied = {k: v for k, v in copied.items() if k not in fields_to_remove}
+                copied.update(fields_to_add)
                 bulk_parent_dicts.append(copied)
 
-        total_preproc_time = 0.001 * RequestMetricsStore.for_request().stop("add_documents.processing_before_opensearch")
-        logger.debug(f"      add_documents pre-processing: took {(total_preproc_time):.3f}s total for {batch_size} docs, "
-                    f"for an average of {(total_preproc_time / batch_size):.3f}s per doc.")
+        total_preproc_time = 0.001 * RequestMetricsStore.for_request().stop(
+            "add_documents.processing_before_opensearch")
+        logger.debug(
+            f"      add_documents pre-processing: took {(total_preproc_time):.3f}s total for {batch_size} docs, "
+            f"for an average of {(total_preproc_time / batch_size):.3f}s per doc.")
 
         logger.debug(f"          add_documents vectorise: took {(total_vectorise_time):.3f}s for {batch_size} docs, "
-                    f"for an average of {(total_vectorise_time / batch_size):.3f}s per doc.")
+                     f"for an average of {(total_vectorise_time / batch_size):.3f}s per doc.")
 
         if bulk_parent_dicts:
-            # the HttpRequest wrapper handles error logic
-            update_mapping_response = backend.add_customer_field_properties(
-                config=config, index_name=add_docs_params.index_name, customer_field_names=new_fields,
-                multimodal_combination_fields=new_obj_fields)
-
             # ADD DOCS TIMER-LOGGER (5)
             start_time_5 = timer()
             with RequestMetricsStore.for_request().time("add_documents.opensearch._bulk"):
-                serialised_body = utils.dicts_to_jsonl(bulk_parent_dicts)
-                index_parent_response = HttpRequests(config).post(
-                    path="_bulk", body=serialised_body)
-            RequestMetricsStore.for_request().add_time("add_documents.opensearch._bulk.internal", float(index_parent_response["took"]))
+                # serialised_body = utils.dicts_to_jsonl(bulk_parent_dicts)
+                vespa_docs = [
+                    VespaDocument(**TypedVespaIndex.to_vespa_document(doc, marqo_index))
+                    for doc in bulk_parent_dicts
+                ]
+                index_responses = vespa_client.feed_batch(vespa_docs, marqo_index.name)
+            RequestMetricsStore.for_request().add_time("add_documents.opensearch._bulk.internal",
+                                                       float(index_parent_response["took"]))
 
             end_time_5 = timer()
             total_http_time = end_time_5 - start_time_5
-            total_index_time = index_parent_response["took"] * 0.001
+            # total_index_time = index_parent_response["took"] * 0.001
             logger.debug(
                 f"      add_documents roundtrip: took {(total_http_time):.3f}s to send {batch_size} docs (roundtrip) to Marqo-os, "
                 f"for an average of {(total_http_time / batch_size):.3f}s per doc.")
 
-            logger.debug(
-                f"          add_documents Marqo-os index: took {(total_index_time):.3f}s for Marqo-os to index {batch_size} docs, "
-                f"for an average of {(total_index_time / batch_size):.3f}s per doc.")
+            # logger.debug(
+            #     f"          add_documents Marqo-os index: took {(total_index_time):.3f}s for Marqo-os to index {batch_size} docs, "
+            #     f"for an average of {(total_index_time / batch_size):.3f}s per doc.")
         else:
             index_parent_response = None
 
         with RequestMetricsStore.for_request().time("add_documents.postprocess"):
-            if add_docs_params.auto_refresh:
-                HttpRequests(config).post(path=F"{add_docs_params.index_name}/_refresh")
-
             t1 = timer()
 
             def translate_add_doc_response(response: Optional[dict], time_diff: float) -> dict:
@@ -707,8 +716,8 @@ def add_documents(config: Config, add_docs_params: AddDocsParams):
                 result_dict["items"] = new_items
                 return result_dict
 
-            return translate_add_doc_response(response=index_parent_response, time_diff=t1 - t0)
-
+            # return translate_add_doc_response(response=index_parent_response, time_diff=t1 - t0)
+            return index_responses
 
 def get_document_by_id(
         config: Config, index_name: str, document_id: str, show_vectors: bool = False):
@@ -894,12 +903,14 @@ def bulk_search(query: BulkSearchQuery, marqo_config: config.Config, verbose: in
     else:
         selected_device = device
 
-    tensor_queries: Dict[int, BulkSearchQueryEntity] = dict(filter(lambda e: e[1].searchMethod == SearchMethod.TENSOR, enumerate(query.queries)))
-    lexical_queries: Dict[int, BulkSearchQueryEntity] = dict(filter(lambda e: e[1].searchMethod == SearchMethod.LEXICAL, enumerate(query.queries)))
+    tensor_queries: Dict[int, BulkSearchQueryEntity] = dict(
+        filter(lambda e: e[1].searchMethod == SearchMethod.TENSOR, enumerate(query.queries)))
+    lexical_queries: Dict[int, BulkSearchQueryEntity] = dict(
+        filter(lambda e: e[1].searchMethod == SearchMethod.LEXICAL, enumerate(query.queries)))
 
     tensor_search_results = dict(zip(tensor_queries.keys(), _bulk_vector_text_search(
-            marqo_config, list(tensor_queries.values()), device=selected_device,
-        )))
+        marqo_config, list(tensor_queries.values()), device=selected_device,
+    )))
 
     # TODO: combine lexical + tensor queries into /_msearch
     lexical_search_results = dict(zip(lexical_queries.keys(), [_lexical_search(
@@ -935,15 +946,18 @@ def bulk_search(query: BulkSearchQuery, marqo_config: config.Config, verbose: in
     }
 
 
-def rerank_query(query: BulkSearchQueryEntity, result: Dict[str, Any], reranker: Union[str, Dict], device: str, num_highlights: int):
+def rerank_query(query: BulkSearchQueryEntity, result: Dict[str, Any], reranker: Union[str, Dict], device: str,
+                 num_highlights: int):
     if query.searchableAttributes is None:
-        raise errors.InvalidArgError(f"searchable_attributes cannot be None when re-ranking. Specify which fields to search and rerank over.")
+        raise errors.InvalidArgError(
+            f"searchable_attributes cannot be None when re-ranking. Specify which fields to search and rerank over.")
     try:
         start_rerank_time = timer()
         rerank.rerank_search_results(search_result=result, query=query.q,
                                      model_name=reranker, device=device,
                                      searchable_attributes=query.searchableAttributes, num_highlights=num_highlights)
-        logger.debug(f"search ({query.searchMethod.lower()}) reranking using {reranker}: took {(timer() - start_rerank_time):.3f}s to rerank results.")
+        logger.debug(
+            f"search ({query.searchMethod.lower()}) reranking using {reranker}: took {(timer() - start_rerank_time):.3f}s to rerank results.")
     except Exception as e:
         raise errors.BadRequestError(f"reranking failure due to {str(e)}")
 
@@ -1199,7 +1213,8 @@ def _lexical_search(
     start_search_http_time = timer()
     with RequestMetricsStore.for_request().time("search.opensearch._search"):
         search_res = HttpRequests(config).get(path=f"{index_name}/_search", body=body)
-    RequestMetricsStore.for_request().add_time("search.opensearch._search.internal", search_res["took"] * 0.001) # internal, not round trip time
+    RequestMetricsStore.for_request().add_time("search.opensearch._search.internal",
+                                               search_res["took"] * 0.001)  # internal, not round trip time
 
     end_search_http_time = timer()
     total_search_http_time = end_search_http_time - start_search_http_time
@@ -1279,20 +1294,21 @@ def construct_msearch_body_elements(searchableAttributes: List[str], offset: int
             )
             if (filter_string is not None) or (searchableAttributes is not None):
                 (search_query["query"]["function_score"]
-                    ["query"]["nested"]
-                    ["query"]["function_score"]
-                    ["query"]["knn"]
-                    [f"{TensorField.chunks}.{TensorField.marqo_knn_field}"]["filter"]) = {
-                        "query_string": {"query": f"{filter_for_opensearch}"}
-                    }
+                ["query"]["nested"]
+                ["query"]["function_score"]
+                ["query"]["knn"]
+                [f"{TensorField.chunks}.{TensorField.marqo_knn_field}"]["filter"]) = {
+                    "query_string": {"query": f"{filter_for_opensearch}"}
+                }
         else:
-            search_query = _create_normal_tensor_search_query(result_count, offset, TensorField.marqo_knn_field, query_vector)
+            search_query = _create_normal_tensor_search_query(result_count, offset, TensorField.marqo_knn_field,
+                                                              query_vector)
             if (filter_string is not None) or (searchableAttributes is not None):
                 (search_query["query"]["nested"]
-                    ["query"]["knn"]
-                    [f"{TensorField.chunks}.{TensorField.marqo_knn_field}"]["filter"]) = {
-                        "query_string": {"query": f"{filter_for_opensearch}"}
-                    }
+                ["query"]["knn"]
+                [f"{TensorField.chunks}.{TensorField.marqo_knn_field}"]["filter"]) = {
+                    "query_string": {"query": f"{filter_for_opensearch}"}
+                }
 
         if attributes_to_retrieve is not None:
             search_query["_source"] = {"include": attributes_to_retrieve} if len(attributes_to_retrieve) > 0 else False
@@ -1311,13 +1327,15 @@ def bulk_msearch(config: Config, body: List[Dict]) -> List[Dict]:
         with RequestMetricsStore.for_request().time("search.opensearch._msearch"):
             serialised_search_body = utils.dicts_to_jsonl(body)
             response = HttpRequests(config).get(path=F"_msearch", body=serialised_search_body)
-        RequestMetricsStore.for_request().add_time("search.opensearch._msearch.internal", float(response["took"])) # internal, not round trip time
+        RequestMetricsStore.for_request().add_time("search.opensearch._msearch.internal",
+                                                   float(response["took"]))  # internal, not round trip time
 
         end_search_http_time = timer()
         total_search_http_time = end_search_http_time - start_search_http_time
         total_os_process_time = response["took"] * 0.001
         num_responses = len(response["responses"])
-        logger.debug(f"search (tensor) roundtrip: took {total_search_http_time:.3f}s to send {num_responses} search queries (roundtrip) to Marqo-os.")
+        logger.debug(
+            f"search (tensor) roundtrip: took {total_search_http_time:.3f}s to send {num_responses} search queries (roundtrip) to Marqo-os.")
 
         responses = [r['hits']['hits'] for r in response["responses"]]
 
@@ -1328,17 +1346,20 @@ def bulk_msearch(config: Config, body: List[Dict]) -> List[Dict]:
             root_cause_type: Optional[str] = response["responses"][0]["error"]["root_cause"][0].get("type")
 
             if "index.max_result_window" in root_cause_reason:
-                raise errors.IllegalRequestedDocCount("Marqo-OS rejected the response due to too many requested results. Try reducing the query's limit parameter") from e
+                raise errors.IllegalRequestedDocCount(
+                    "Marqo-OS rejected the response due to too many requested results. Try reducing the query's limit parameter") from e
             elif 'parse_exception' in root_cause_reason:
                 raise errors.InvalidArgError("Syntax error, could not parse filter string") from e
-            elif  root_cause_type == 'query_shard_exception' and root_cause_reason.startswith("Failed to parse query"):
+            elif root_cause_type == 'query_shard_exception' and root_cause_reason.startswith("Failed to parse query"):
                 raise errors.InvalidArgError("Syntax error, could not parse filter string") from e
             raise errors.BackendCommunicationError(f"Error communicating with Marqo-OS backend:\n{response}")
         except (KeyError, IndexError):
             raise e
 
-    logger.debug(f"  search (tensor) Marqo-os processing time: took {total_os_process_time:.3f}s for Marqo-os to execute the search.")
+    logger.debug(
+        f"  search (tensor) Marqo-os processing time: took {total_os_process_time:.3f}s for Marqo-os to execute the search.")
     return responses
+
 
 def gather_documents_from_response(resp: List[List[Dict[str, Any]]]) -> Dict[str, Any]:
     """
@@ -1419,7 +1440,8 @@ def assign_query_to_vector_job(
     return ptrs
 
 
-def create_vector_jobs(queries: List[BulkSearchQueryEntity], config: Config, device: str) -> Tuple[Dict[Qidx, List[VectorisedJobPointer]], Dict[JHash, VectorisedJobs]]:
+def create_vector_jobs(queries: List[BulkSearchQueryEntity], config: Config, device: str) -> Tuple[
+    Dict[Qidx, List[VectorisedJobPointer]], Dict[JHash, VectorisedJobs]]:
     """
         For each query:
             - Find what needs to be vectorised
@@ -1509,7 +1531,8 @@ def get_query_vectors_from_jobs(
                     possible_jobs=qidx_to_job[qidx],
                     jobs=jobs,
                     job_to_vectors=job_to_vectors,
-                    treat_urls_as_images=index_info.index_settings[NsField.index_defaults][NsField.treat_urls_and_pointers_as_images],
+                    treat_urls_as_images=index_info.index_settings[NsField.index_defaults][
+                        NsField.treat_urls_and_pointers_as_images],
                     content=content),
                  weight,
                  content
@@ -1540,8 +1563,9 @@ def get_query_vectors_from_jobs(
             result[qidx] = get_content_vector(
                 possible_jobs=qidx_to_job.get(qidx, []),
                 jobs=jobs,
-                job_to_vectors= job_to_vectors,
-                treat_urls_as_images=index_info.index_settings[NsField.index_defaults][NsField.treat_urls_and_pointers_as_images],
+                job_to_vectors=job_to_vectors,
+                treat_urls_as_images=index_info.index_settings[NsField.index_defaults][
+                    NsField.treat_urls_and_pointers_as_images],
                 content=q.q
             )
     return result
@@ -1581,11 +1605,14 @@ def create_empty_query_response(queries: List[BulkSearchQueryEntity]) -> List[Di
         )
     )
 
-def run_vectorise_pipeline(config: Config, queries: List[BulkSearchQueryEntity], device: Union[Device, str]) -> Dict[Qidx, List[float]]:
+
+def run_vectorise_pipeline(config: Config, queries: List[BulkSearchQueryEntity], device: Union[Device, str]) -> Dict[
+    Qidx, List[float]]:
     """Run the query vectorisation process"""
     # 1. Pre-process inputs ready for s2_inference.vectorise
     # we can still use qidx_to_job. But the jobs structure may need to be different
-    vector_jobs_tuple: Tuple[Dict[Qidx, List[VectorisedJobPointer]], Dict[JHash, VectorisedJobs]] = create_vector_jobs(queries, config, device)
+    vector_jobs_tuple: Tuple[Dict[Qidx, List[VectorisedJobPointer]], Dict[JHash, VectorisedJobs]] = create_vector_jobs(
+        queries, config, device)
 
     qidx_to_jobs, jobs = vector_jobs_tuple
 
@@ -1599,6 +1626,7 @@ def run_vectorise_pipeline(config: Config, queries: List[BulkSearchQueryEntity],
         queries, qidx_to_jobs, job_ptr_to_vectors, config, jobs
     )
     return qidx_to_vectors
+
 
 def _bulk_vector_text_search(config: Config, queries: List[BulkSearchQueryEntity], device: str = None) -> List[Dict]:
     """Resolve a batch of search queries in parallel.
@@ -1620,18 +1648,21 @@ def _bulk_vector_text_search(config: Config, queries: List[BulkSearchQueryEntity
         raise errors.InternalError("_bulk_vector_text_search cannot be called without `device`!")
 
     with RequestMetricsStore.for_request().time("bulk_search.vector.processing_before_opensearch",
-        lambda t : logger.debug(f"bulk search (tensor) pre-processing: took {t:.3f}ms")
-    ):
+                                                lambda t: logger.debug(
+                                                    f"bulk search (tensor) pre-processing: took {t:.3f}ms")
+                                                ):
 
         with RequestMetricsStore.for_request().time(f"bulk_search.vector_inference_full_pipeline"):
             qidx_to_vectors: Dict[Qidx, List[float]] = run_vectorise_pipeline(config, queries, device)
 
         ## 4. Create msearch request bodies and combine to aggregate.
         query_to_body_parts: Dict[Qidx, List[Dict]] = dict()
-        query_to_body_count: Dict[Qidx, int] = dict() # Keep track of count, so we can separate after msearch call.
+        query_to_body_count: Dict[Qidx, int] = dict()  # Keep track of count, so we can separate after msearch call.
         for qidx, q in enumerate(queries):
             index_info = get_index_info(config=config, index_name=q.index)
-            body = construct_msearch_body_elements(q.searchableAttributes, q.offset, q.filter, index_info, q.limit, qidx_to_vectors[qidx], q.attributesToRetrieve, q.index, q.scoreModifiers)
+            body = construct_msearch_body_elements(q.searchableAttributes, q.offset, q.filter, index_info, q.limit,
+                                                   qidx_to_vectors[qidx], q.attributesToRetrieve, q.index,
+                                                   q.scoreModifiers)
             query_to_body_parts[qidx] = body
             query_to_body_count[qidx] = len(body)
 
@@ -1645,13 +1676,15 @@ def _bulk_vector_text_search(config: Config, queries: List[BulkSearchQueryEntity
     responses = bulk_msearch(config, aggregate_body)
 
     with RequestMetricsStore.for_request().time("bulk_search.vector.postprocess",
-        lambda t : logger.debug(f"bulk search (tensor) post-processing: took {t:.3f}ms")
-    ):
+                                                lambda t: logger.debug(
+                                                    f"bulk search (tensor) post-processing: took {t:.3f}ms")
+                                                ):
         # 6. Get documents back to each query, perform "gather" operation
         return create_bulk_search_response(queries, query_to_body_count, responses)
 
 
-def create_bulk_search_response(queries: List[BulkSearchQueryEntity], query_to_body_count: Dict[Qidx, int], responses) -> List[Dict]:
+def create_bulk_search_response(queries: List[BulkSearchQueryEntity], query_to_body_count: Dict[Qidx, int],
+                                responses) -> List[Dict]:
     """
         Create Marqo search responses by extracting the appropriate elements from the batched /_msearch response. Also handles:
             - Boosting score (optional)
@@ -1678,6 +1711,7 @@ def create_bulk_search_response(queries: List[BulkSearchQueryEntity], query_to_b
         )
 
     return results
+
 
 def _vector_text_search(
         config: Config, index_name: str, query: Union[str, dict], result_count: int = 5, offset: int = 0,
@@ -1736,21 +1770,26 @@ def _vector_text_search(
         raise errors.IndexNotFoundError(message="Tried to search a non-existent index: {}".format(index_name))
 
     queries = [BulkSearchQueryEntity(
-        q=query, searchableAttributes=searchable_attributes,searchMethod=SearchMethod.TENSOR, limit=result_count, offset=offset, showHighlights=False, filter=filter_string, attributesToRetrieve=attributes_to_retrieve, boost=boost, image_download_headers=image_download_headers, context=context, scoreModifiers=score_modifiers, index=index_name, modelAuth=model_auth
+        q=query, searchableAttributes=searchable_attributes, searchMethod=SearchMethod.TENSOR, limit=result_count,
+        offset=offset, showHighlights=False, filter=filter_string, attributesToRetrieve=attributes_to_retrieve,
+        boost=boost, image_download_headers=image_download_headers, context=context, scoreModifiers=score_modifiers,
+        index=index_name, modelAuth=model_auth
     )]
     with RequestMetricsStore.for_request().time(f"search.vector_inference_full_pipeline"):
         qidx_to_vectors: Dict[Qidx, List[float]] = run_vectorise_pipeline(config, queries, device)
     vectorised_text = list(qidx_to_vectors.values())[0]
 
     body = construct_msearch_body_elements(
-        searchable_attributes, offset, filter_string, index_info, result_count, vectorised_text, attributes_to_retrieve, index_name, score_modifiers
+        searchable_attributes, offset, filter_string, index_info, result_count, vectorised_text, attributes_to_retrieve,
+        index_name, score_modifiers
     )
 
     if verbose:
         _vector_text_search_query_verbose(verbose=verbose, body=body)
 
     total_preprocess_time = RequestMetricsStore.for_request().stop("search.vector.processing_before_opensearch")
-    logger.debug(f"search (tensor) pre-processing: took {(total_preprocess_time):.3f}ms to vectorize and process query.")
+    logger.debug(
+        f"search (tensor) pre-processing: took {(total_preprocess_time):.3f}ms to vectorize and process query.")
 
     # SEARCH TIMER-LOGGER (roundtrip)
     responses = bulk_msearch(config, body)
@@ -1778,6 +1817,7 @@ def _vector_text_search(
         f"search (tensor) post-processing: took {(total_postprocess_time):.3f}ms to sort and format {len(completely_sorted)} results from Marqo-os.")
     return res
 
+
 def _format_ordered_docs_simple(ordered_docs_w_chunks: List[dict], result_count: int) -> dict:
     """Only one highlight is returned
     Args:
@@ -1800,6 +1840,7 @@ def _format_ordered_docs_simple(ordered_docs_w_chunks: List[dict], result_count:
         simple_results.append(cleaned)
     return {"hits": simple_results[:result_count]}
 
+
 def boost_score(docs: dict, boosters: dict, searchable_attributes) -> dict:
     """ re-weighs the scores of individual fields
         Args:
@@ -1811,10 +1852,10 @@ def boost_score(docs: dict, boosters: dict, searchable_attributes) -> dict:
     if searchable_attributes and boosters:
         if not set(boosters).issubset(set(searchable_attributes)):
             raise errors.InvalidArgError(
-        "Boost fieldnames must be a subset of searchable attributes. "
-        f"\nSearchable attributes: {searchable_attributes}"
-        f"\nBoost: {boosters}"
-        )
+                "Boost fieldnames must be a subset of searchable attributes. "
+                f"\nSearchable attributes: {searchable_attributes}"
+                f"\nBoost: {boosters}"
+            )
 
     for doc_id in list(to_be_boosted.keys()):
         for chunk in to_be_boosted[doc_id]["chunks"]:
@@ -1858,6 +1899,7 @@ def check_health(config: Config) -> dict:
     Please use check_index_health instead.
     """
     return generate_heath_check_response(config)
+
 
 def delete_index(config: Config, index_name):
     res = HttpRequests(config).delete(path=index_name)
@@ -1930,7 +1972,7 @@ def get_cuda_info() -> dict:
 
 def vectorise_multimodal_combination_field(
         field: str, multimodal_object: Dict[str, dict], doc: dict, doc_index: int,
-        doc_id:str, device:str, index_info, image_repo, field_map:dict,
+        doc_id: str, device: str, index_info, image_repo, field_map: dict,
         model_auth: Optional[ModelAuth] = None
 ):
     '''
@@ -1986,8 +2028,9 @@ def vectorise_multimodal_combination_field(
     if infer_if_image is False:
         text_field_names = list(multimodal_object.keys())
         text_content_to_vectorise = list(multimodal_object.values())
-        new_fields_from_multimodal_combination =set([(sub_field_name, _infer_opensearch_data_type(sub_content)) for sub_field_name
-        , sub_content in multimodal_object.items()])
+        new_fields_from_multimodal_combination = set(
+            [(sub_field_name, _infer_opensearch_data_type(sub_content)) for sub_field_name
+            , sub_content in multimodal_object.items()])
     else:
         for sub_field_name, sub_content in multimodal_object.items():
             if isinstance(sub_content, str) and not _is_image(sub_content):
@@ -1996,7 +2039,7 @@ def vectorise_multimodal_combination_field(
             else:
                 try:
                     if isinstance(sub_content, str) and index_info.index_settings[NsField.index_defaults][
-                            NsField.treat_urls_and_pointers_as_images]:
+                        NsField.treat_urls_and_pointers_as_images]:
                         if not isinstance(image_repo[sub_content], Exception):
                             image_data = image_repo[sub_content]
                         else:
@@ -2054,7 +2097,7 @@ def vectorise_multimodal_combination_field(
         image_err = errors.InvalidArgError(message=f'Could not process given image: {multimodal_object_copy}')
         unsuccessful_doc_to_append = \
             (doc_index, {'_id': doc_id, 'error': image_err.message, 'status': int(image_err.status_code),
-                 'code': image_err.code})
+                         'code': image_err.code})
 
         return combo_chunk, combo_document_is_valid, unsuccessful_doc_to_append, combo_vectorise_time_to_add, new_fields_from_multimodal_combination
 
@@ -2062,9 +2105,12 @@ def vectorise_multimodal_combination_field(
     vectors_list = text_vectors + image_vectors
 
     if not len(sub_field_name_list) == len(vectors_list):
-        raise errors.BatchInferenceSizeError(message=f"Batch inference size does not match content for multimodal field {field}")
+        raise errors.BatchInferenceSizeError(
+            message=f"Batch inference size does not match content for multimodal field {field}")
 
-    vector_chunk = np.squeeze(np.mean([np.array(vector) * field_map["weights"][sub_field_name] for sub_field_name, vector in zip(sub_field_name_list, vectors_list)], axis=0))
+    vector_chunk = np.squeeze(np.mean(
+        [np.array(vector) * field_map["weights"][sub_field_name] for sub_field_name, vector in
+         zip(sub_field_name_list, vectors_list)], axis=0))
 
     if normalize_embeddings is True:
         vector_chunk = vector_chunk / np.linalg.norm(vector_chunk)
@@ -2078,7 +2124,8 @@ def vectorise_multimodal_combination_field(
     })
     return combo_chunk, combo_document_is_valid, unsuccessful_doc_to_append, combo_vectorise_time_to_add, new_fields_from_multimodal_combination
 
-def _generate_vector_text_search_query_for_verbose_one(original_body:List[Dict[str, Any]]) -> None:
+
+def _generate_vector_text_search_query_for_verbose_one(original_body: List[Dict[str, Any]]) -> None:
     """Generate a simplified version of the query body for verbose=1 mode. """
     readable_body = copy.deepcopy(original_body)
     for i, q in enumerate(readable_body):
@@ -2151,7 +2198,8 @@ def _create_normal_tensor_search_query(result_count, offset, vector_field, vecto
     return search_query
 
 
-def _create_score_modifiers_tensor_search_query(score_modifiers, result_count, offset, vector_field, vectorised_text) -> dict:
+def _create_score_modifiers_tensor_search_query(score_modifiers, result_count, offset, vector_field,
+                                                vectorised_text) -> dict:
     script_score = score_modifiers.to_painless_script()
     search_query = {
         "size": result_count,
